type-inference: Infer Array with free variable generic
When we see an array literal and have no assignment information
we infer that it is Array<T>, aka an unbound generic array with
a free variable T.

We later change the type of the variable based on the first usage.
In this case the array is late bound to a single concrete type.